backend/requirement.txt
blinker==1.8.2
boto3==1.34.150
botocore==1.34.150
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
urllib3==1.26.19
Werkzeug==3.0.3
zipp==3.19.2
